add files response handler
