+++ packages/paginated-table/src/wrapper/components/AddAllFiles/AddAllView.js
@@ -4,6 +4,7 @@ import clsx from 'clsx';
 import { getFilesID } from '../../WrapperService';
 import AddToCartDialogView from '../AddToCartDialog/AddToCartDialogView';
 import ToolTipView from '../TooltipView';
+import addFilesResponseHandler from '../util';
 
 const AddAllFilesComponent = (props) => {
   const {
@@ -38,15 +39,7 @@ const AddAllFilesComponent = (props) => {
     fileIds().then((response) => {
       const data = response[responseKeys[0]];
       if (data && data.length > 0) {
-        const isArray = Array.isArray(data[0][responseKeys[1]]);
-        const ids = data.reduce((acc, id) => {
-          if (id && id[responseKeys[1]]) {
-            // if object convert to array
-            const items = isArray ? id[responseKeys[1]] : [id[responseKeys[1]]];
-            acc.push(...items);
-          }
-          return acc;
-        }, []);
+        const ids = addFilesResponseHandler(response, responseKeys);
         if (ids.length > 1000) {
           setAlterDisplay(true);
         } else {
