+++ packages/paginated-table/src/wrapper/components/AddSelectedFiles/AddSelectedFilesController.js
@@ -6,6 +6,7 @@ import { TableContext } from '../../../table/ContextProvider';
 import { onRowSeclect } from '../../../table/state/Actions';
 import AddSelectedFileComponent from './AddSelectedFilesView';
 import { getFilesID } from '../../WrapperService';
+import addFilesResponseHandler from '../util';
 
 const AddSelectedFilesController = (props) => {
   const {
@@ -26,8 +27,10 @@ const AddSelectedFilesController = (props) => {
     selectedRows = [],
     dispatch,
   } = context;
-  const variables = {};
-  variables[dataKey] = selectedRows;
+  const variables = {
+    first: 10000,
+    [dataKey]: selectedRows,
+  };
   // add selected files id
   const addSelectedFiles = () => {
     const fileIds = getFilesID({
@@ -38,7 +41,7 @@ const AddSelectedFilesController = (props) => {
     });
 
     fileIds().then((response) => {
-      const ids = response[responseKeys[0]] || [];
+      const ids = addFilesResponseHandler(response, responseKeys);
       if (ids.length >= 1000) {
         setAlterDisplay(true);
       } else {
